Latest update, using base model template to fix Django.
