+++ bakhanapp/models.py
@@ -145,17 +145,11 @@ class Chapter(models.Model):
 
 class Chapter_Mineduc(models.Model):
     id_chapter_mineduc = models.AutoField(primary_key=True)
-    id_subject = models.ForeignKey('Subject')
-    level = models.IntegerField()
-    year = models.IntegerField()
-    additional = models.TextField(blank=True, null=True)
+    index = models.IntegerField(blank=True, null=True)
+    name = models.CharField(unique=True, max_length=50, blank=True, null=True)
 
     class Meta:
         db_table = 'bakhanapp_chapter_mineduc'
-        unique_together = (('level','year','id_subject'),)
-
-    def __unicode__(self): # __unicode__ on Python 2
-        return self.level
 
 
 class Class(models.Model):
@@ -188,14 +182,10 @@ class Class_Subject(models.Model):
     id_class = models.ForeignKey(Class)
     id_subject_name = models.ForeignKey('Subject')
     kaid_teacher = models.ForeignKey('Teacher')
-    curriculum = models.ForeignKey(Chapter_Mineduc, blank=True, null=True)
 
     class Meta:
         db_table = 'bakhanapp_class_subject'
         unique_together = (('id_class', 'id_subject_name'),)
-
-    def __unicode__(self): # __unicode__ on Python 2
-        return self.name_spanish
 
 
 #Al parecer este modelo no es usado ya que solo existia en la BD y no en Django(?)
@@ -282,19 +272,20 @@ class Institution(models.Model):
 
 class Planning(models.Model):
     id_planning = models.AutoField(primary_key=True)
-    class_subject = models.ForeignKey(Class_Subject)
-    class_subtopic = models.ForeignKey('Subtopic_Mineduc')
-    class_date = models.DateField(blank=True, null=True)
-    minutes = models.IntegerField(blank=True, null=True)
-    status = models.BooleanField(default= False)
-    desc_inicio = models.TextField(blank=True, null=True)
-    desc_cierre = models.TextField(blank=True, null=True)
-    share_class = models.NullBooleanField(default=False)
-    class_name = models.TextField(blank=False, null=False)
+    curso = models.CharField(max_length=20, blank=True, null=True)
+    oa = models.CharField(max_length=500, blank=True, null=True)
+    clase = models.CharField(max_length=50, blank=True, null=True)
+    objetivo = models.CharField(max_length=500, blank=True, null=True)
+    inicio = models.CharField(max_length=500, blank=True, null=True)
+    ejerciciokhan = models.CharField(max_length=500, blank=True, null=True)
+    videokhan = models.CharField(max_length=500, blank=True, null=True)
+    descripcion = models.CharField(max_length=500, blank=True, null=True)
+    cierre = models.CharField(max_length=500, blank=True, null=True)
+    teacher = models.ForeignKey('Teacher', blank=True, null=True)
 
     class Meta:
         db_table = 'bakhanapp_planning'
-        unique_together = (('class_name', 'class_subject'),)
+        unique_together = (('curso', 'oa'),)
 
 
 class Related_Video_Exercise(models.Model):
@@ -367,13 +358,11 @@ class Skill_Log(models.Model):
 
 class Skill_Planning(models.Model):
     id_skill_planning = models.AutoField(primary_key=True)
-    id_planning = models.ForeignKey(Planning)
-    id_skill = models.ForeignKey(Skill)
-    id_subtopic = models.ForeignKey("Subtopic_Skill")
+    id_planning = models.ForeignKey(Planning, blank=True, null=True)
+    id_skill = models.ForeignKey(Skill, blank=True, null=True)
 
     class Meta:
         db_table = 'bakhanapp_skill_planning'
-        unique_together = (('id_planning','id_skill'),('id_planning','id_subtopic'),)
 
 
 class Skill_Progress(models.Model):
@@ -479,14 +468,14 @@ class Subtopic(models.Model):
 
 class Subtopic_Mineduc(models.Model):
     id_subtopic_mineduc = models.AutoField(primary_key=True)
-    id_topic = models.ForeignKey('Topic_Mineduc')
-    index = models.IntegerField()
-    description = models.TextField(blank=True, null=True)
-    summary = models.TextField(blank=True, null=True)
+    name = models.CharField(max_length=50, blank=True, null=True)
+    id_topic = models.ForeignKey('Topic_Mineduc', blank=True, null=True)
+    ae_oe = models.CharField(db_column='AE_OE', max_length=500, blank=True, null=True)  # Field name made lowercase.
+    index = models.IntegerField(blank=True, null=True)
 
     class Meta:
         db_table = 'bakhanapp_subtopic_mineduc'
-        unique_together = (('index', 'id_topic'),)
+        unique_together = (('name', 'id_topic'),)
 
 
 class Subtopic_Skill(models.Model):
@@ -557,13 +546,14 @@ class Topic(models.Model):
 
 class Topic_Mineduc(models.Model):
     id_topic_mineduc = models.AutoField(primary_key=True)
-    id_chapter = models.ForeignKey(Chapter_Mineduc)
-    index = models.IntegerField()
-    descripcion_topic = models.TextField(blank=True, null=True)
+    name = models.CharField(max_length=50, blank=True, null=True)
+    id_chapter = models.ForeignKey(Chapter_Mineduc, blank=True, null=True)
+    index = models.IntegerField(blank=True, null=True)
+    descripcion_topic = models.CharField(max_length=250, blank=True, null=True)
 
     class Meta:
         db_table = 'bakhanapp_topic_mineduc'
-        unique_together = (('index','id_chapter'),)
+        unique_together = (('name', 'id_chapter'),)
 
 
 class Tutor(models.Model):
@@ -604,13 +594,11 @@ class Video(models.Model):
 
 class Video_Planning(models.Model):
     id_video_planning = models.AutoField(primary_key=True)
-    id_planning = models.ForeignKey(Planning)
-    id_video = models.ForeignKey(Video)
-    id_subtopic = models.ForeignKey(Subtopic_Video)
+    id_planning = models.ForeignKey(Planning, blank=True, null=True)
+    id_video = models.ForeignKey(Video, blank=True, null=True)
 
     class Meta:
         db_table = 'bakhanapp_video_planning'
-        unique_together = (('id_planning','id_video'),('id_planning','id_subtopic'),)
 
 
 class Video_Playing(models.Model):
@@ -626,3 +614,4 @@ class Video_Playing(models.Model):
     class Meta:
         db_table = 'bakhanapp_video_playing'
         unique_together = (('date', 'id_video_name', 'kaid_student'),)
+
